refactor prout and cloudformation logic
